backend/pyproject.toml
[project]
name = "telemedicine-backend"
version = "0.1.0"
description = "Secure FastAPI backend for TeleMedicine"
authors = [{ name = "TeleMedicine Team" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard]>=0.116.1,<0.117.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "alembic>=1.16.5,<2.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "pyjwt[crypto]>=2.10.1,<3.0.0",
    "bcrypt>=4.3.0,<5.0.0",
    "python-multipart>=0.0.20,<0.0.21",
    "gunicorn>=23.0.0,<24.0.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
ruff = "^0.12.11"
aiosqlite = "^0.21.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
